src/requirements.txt
aiofiles==0.6.0
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
aiostalk==1.1
anyio==3.6.1
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==1.0.0a4
click==8.1.3
commonmark==0.9.1
crashtest==0.3.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
executing==1.1.1
factory-boy==3.2.1
Faker==15.1.1
fastapi==0.85.1
fastapi-cache2==0.1.9
frozenlist==1.3.1
greenstalk==2.0.0
h11==0.14.0
hiredis==2.0.0
httpie==3.2.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
multidict==6.0.2
numpy==1.23.4
packaging==21.3
pandas==1.5.0
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pptree==3.1
praw==7.6.0
prawcore==2.3.0
prompt-toolkit==3.0.31
psaw==0.1.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
pylev==1.4.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-ulid==1.1.0
pytz==2022.4
pytzdata==2020.1
redis==4.3.4
redis-om==0.0.27
requests==2.28.1
requests-toolbelt==0.10.0
rich==12.6.0
six==1.16.0
sniffio==1.3.0
stack-data==0.5.1
starlette==0.20.4
traitlets==5.4.0
types-redis==4.3.21.2
types-six==1.16.21
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
websocket-client==1.4.1
wrapt==1.14.1
yarl==1.8.1
